[cpp] Do not write hxcpp path if not compiling
When building the haxelib binary, no haxelib executable is available
in PATH yet, so the haxelib call is a problem.

With `-D no-compilation`, haxe doesn't use/run hxcpp at all so this path
is not very meaningful. The user may compile manually with a completely
different version of hxcpp.

Either way, this value is completely ignored by hxcpp:
https://github.com/HaxeFoundation/hxcpp/blob/08f88ff314eeaa165dd76272f92c8030cc879ec2/tools/hxcpp/BuildTool.hx#L1729-L1732